+++ src/configs/jest.ts
@@ -19,9 +19,13 @@ type AllJestMatchers = PrettifyShallow<keyof ReturnType<JestExpect> | keyof Asym
 
 export const generateDefaultTestFiles = <T extends string>(extensions: T) => [
   `**/*.spec.${extensions}` as const, // GitHub: 2.3M .ts files as of 2024-12-08 (https://github.com/search?q=path%3A**%2F*.spec.ts&type=code&query=path%3A%2F**%2F__tests__%2F**%2F*.ts)
+  `**/-spec.${extensions}` as const, // 165k
+  `**/_spec.${extensions}` as const, // 40k
+
   `**/*.test.${extensions}` as const, // 1.9M
-  `__tests__/**/*.${extensions}` as const, // 155k
-  `__test__/**/*.${extensions}` as const, // 14k
+
+  `**/__tests__/**/*.${extensions}` as const, // 155k
+  `**/__test__/**/*.${extensions}` as const, // 14k
 ];
 
 export interface JestEslintConfigOptions extends ConfigSharedOptions<'jest'> {