+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Dependencies
 - `typescript-eslint`: 8.20.0 -> 8.23.0
 - `@eslint/markdown`: 6.2.1 -> 6.2.2
+- `@stylistic/eslint-plugin`: 2.13.0 -> 3.1.0
 
 ## 0.4.1