+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - `typescript-eslint`: 8.20.0 -> 8.23.0
 - `@eslint/markdown`: 6.2.1 -> 6.2.2
 - `@stylistic/eslint-plugin`: 2.13.0 -> 3.1.0
+- `@vitest/eslint-plugin`: 1.1.25 -> 1.1.28
+  - Enabled a new [`require-mock-type-parameters`](https://github.com/vitest-dev/eslint-plugin-vitest/blob/main/docs/rules/require-mock-type-parameters.md) rule with ⚠️warning severity.
 
 ## 0.4.1