+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - `@stylistic/eslint-plugin`: 2.13.0 -> 3.1.0
 - `@vitest/eslint-plugin`: 1.1.25 -> 1.1.28
   - Enabled a new [`require-mock-type-parameters`](https://github.com/vitest-dev/eslint-plugin-vitest/blob/main/docs/rules/require-mock-type-parameters.md) rule with ⚠️warning severity.
+- `eslint-config-flat-gitignore`: 1.0.0 -> 2.0.0
 
 ## 0.4.1