+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - `jest`/`vitest`: by default include nested `__test(s)__` directories and `[-_].spec.*` files.
+- `sonar`: change `prefer-single-boolean-return` default severity to `warn`.
 
 ### Dependencies
 - `typescript-eslint`: 8.20.0 -> 8.24.0