+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 0.4.2
 
 - New config: `cli` to disable a few rules for files in `bin`, `scripts` and `cli` directories, enabled by default.
 - `jest`/`vitest`: by default include nested `__test(s)__` directories and `[-_].spec.*` files.